+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project follows [Keep a Changelog](https://keepachangelog.com/) and [Semant
 ### Changed
 
 - Added Example warings to JSON output of scan usage.
+- Changed scan statistics name from "Total usages found" to "Total variables".
 
 ### Fixed